requirements.txt
streamlit
streamlit-chat
python-dotenv
azure-ai-ml
azure-identity
pandas